Compute feature-wise absolute difference penalty

Compute feature-wise absolute difference penalty

Compute t-stat absolute difference penalty

colmeans_penalty(x1, x2)

p_transform(penalty)

tstat_penalty(x1, x2, y1, y2)

Arguments

x1

A data matrix

x2

A data matrix

penalty

A vector of penalties to be transformed where the sum is set to the length of the vector

Value

A vector A vector

Examples

n = 10
p = 10
x1 = matrix(rnorm(n * p, mean = 0, sd = 1), nrow = n, ncol = p)
x2 = matrix(rnorm(n * p, mean = 1, sd = 1), nrow = n, ncol = p)
colmeans_penalty(x1, x2)
#>  [1] 0.8236404 1.2599340 1.3849807 1.7834853 0.6810144 0.6335384 0.6399791
#>  [8] 0.2126798 1.0025180 1.5782300
n = 10
p = 10
x1 = matrix(rnorm(n * p, mean = 0, sd = 1), nrow = n, ncol = p)
y1 = factor(rbinom(n, 1, prob = 0.5), levels = c("0", "1"))
x2 = matrix(rnorm(n * p, mean = 1, sd = 1), nrow = n, ncol = p)
y2 = factor(rbinom(n, 1, prob = 0.5), levels = c("0", "1"))
tstat_penalty(x1 = x1, x2 = x2, y1 = y1, y2 = y2)
#>  [1] 1.0477230 1.4307539 0.4106593 0.1466568 0.7676688 0.9818633 0.8596887
#>  [8] 2.2778327 1.0326975 1.0444560