CPOP modelling

CPOP is consisted of three steps. Step 1 is to select features common to two transformed data. Note the input must be pairwise-differences between the original data columns. Step 2 is to select features in constructed models that shared similar characteristics. Step 3 is to construct a final model used for prediction.

cpop_model(
  z1,
  z2,
  y1,
  y2,
  w = NULL,
  n_features = 50,
  n_iter = 20,
  alpha = 1,
  family = "binomial",
  s = "lambda.min",
  cpop2_break = TRUE,
  cpop2_type = "sign",
  cpop2_mag = 1,
  cpop1_method = "normal",
  intercept = FALSE,
  ...
)

Arguments

z1
A data matrix, columns are pairwise-differences between the original data columns.

z2
A data matrix, columns are pairwise-differences between the original data columns. Column names should be identical to z1.

y1
A vector of response variable. Same length as the number of rows of z1.

y2
A vector of response variable. Same length as the number of rows of z2.

w
A vector of weights.

n_features
Breaking the CPOP-Step 1 loop if a certain number of features is reached.

n_iter
Number of iterations in Step 1 and 2.

alpha
The alpha parameter for elastic net models. See the alpha argument in glmnet::glmnet.

family
family of glmnet

s
CV-Lasso lambda

cpop2_break
Should cpop2 loop be broken the first time

cpop2_type
"sign" or "mag"

cpop2_mag
a threshold differential betas are removed

intercept
Default to FALSE

...
Extra parameter settings for cv.glmnet in cpop1

Value

A vector

Examples

data(cpop_data_binary, package = 'CPOP')
attach(cpop_data_binary)
#> The following objects are masked from cpop_data_binary (pos = 3):
#> 
#>     beta, x1, x2, x3, y1, y2, y3
#> The following objects are masked from cpop_data_binary (pos = 4):
#> 
#>     beta, x1, x2, x3, y1, y2, y3
set.seed(1)
z1 = pairwise_col_diff(x1)
z2 = pairwise_col_diff(x2)
cpop_result = cpop_model(z1, z2, y1, y2, alpha = 1, n_features = 10)
#> Absolute colMeans difference will be used as the weights for CPOP
#> Fitting CPOP model using alpha = 1
#> Based on previous alpha, 0 features are kept 
#> CPOP1 - Step 01: Number of selected features: 0 out of 190
#> CPOP1 - Step 02: Number of selected features: 9 out of 190
#> CPOP1 - Step 03: Number of selected features: 16 out of 190
#> 10 features was reached. 
#> A total of 16 features were selected. 
#> Removing sources of collinearity gives 13 features. 
#> 10 features was reached. 
#> A total of 13 features were selected. 
#> CPOP2 - Sign: Step 01: Number of leftover features: 9 out of 13
#> The sign matrix between the two data:
#>     
#>      -1 0 1
#>   -1  0 0 1
#>   0   0 0 0
#>   1   3 0 0
#> CPOP2 - Sign: Step 02: Number of leftover features: 8 out of 13
#> The sign matrix between the two data:
#>     
#>      -1 0 1
#>   -1  0 0 0
#>   0   0 0 0
#>   1   1 0 0
#> CPOP2 - Sign: Step 03: Number of leftover features: 8 out of 13
#> The sign matrix between the two data:
#>     
#>      -1 0 1
#>   -1  0 0 0
#>   0   0 0 0
#>   1   0 0 0
cpop_result
#> $glmnet1
#> 
#> Call:  glmnet::cv.glmnet(x = z1[, cpop2_result], y = y1, family = family,      alpha = 0, intercept = intercept) 
#> 
#> Measure: Binomial Deviance 
#> 
#>      Lambda Index Measure      SE Nonzero
#> min 0.02761   100  0.7775 0.06786       8
#> 1se 0.07000    90  0.8416 0.06030       8
#> 
#> $glmnet2
#> 
#> Call:  glmnet::cv.glmnet(x = z2[, cpop2_result], y = y2, family = family,      alpha = 0, intercept = intercept) 
#> 
#> Measure: Binomial Deviance 
#> 
#>      Lambda Index Measure      SE Nonzero
#> min 0.03676    96  0.9288 0.09282       8
#> 1se 0.25935    75  1.0137 0.05566       8
#> 
#> $feature
#> [1] "X01--X10" "X09--X17" "X11--X14" "X12--X20" "X01--X07" "X01--X15" "X01--X17"
#> [8] "X04--X12"
#> 
#> $coef_tbl
#> # A tibble: 9 × 3
#>   coef_name    coef1    coef2
#>   <chr>        <dbl>    <dbl>
#> 1 (Intercept)  0      0      
#> 2 X01--X10    -0.322 -0.246  
#> 3 X09--X17     0.722  0.521  
#> 4 X11--X14     0.130  0.00292
#> 5 X12--X20     0.404  0.170  
#> 6 X01--X07    -0.437 -0.408  
#> 7 X01--X15    -0.158 -0.334  
#> 8 X01--X17    -0.901 -0.644  
#> 9 X04--X12     0.353  0.431  
#> 
#> $cpop1_features
#>  [1] "X01--X10" "X05--X06" "X05--X09" "X09--X17" "X11--X14" "X12--X20"
#>  [7] "X16--X19" "X01--X07" "X01--X15" "X01--X17" "X04--X12" "X06--X16"
#> [13] "X08--X19"
#> 
#> $step_features
#> # A tibble: 78 × 6
#>    step  coef_model feature_name feature_value ever_selected_f… in_final_featur…
#>    <chr> <chr>      <chr>                <dbl> <lgl>            <lgl>           
#>  1 1     1          (Intercept)         0.232  FALSE            FALSE           
#>  2 1     1          X01--X10           -2.34   TRUE             TRUE            
#>  3 1     1          X03--X07            0.321  FALSE            FALSE           
#>  4 1     1          X04--X20            0.429  TRUE             FALSE           
#>  5 1     1          X05--X06            0.417  TRUE             TRUE            
#>  6 1     1          X05--X09           -0.270  TRUE             TRUE            
#>  7 1     1          X05--X16            0.0627 FALSE            FALSE           
#>  8 1     1          X06--X08            0.175  FALSE            FALSE           
#>  9 1     1          X06--X16            0.417  TRUE             TRUE            
#> 10 1     1          X06--X19            0.0950 FALSE            FALSE           
#> # … with 68 more rows
#> 
#> attr(,"class")
#> [1] "cpop" "list"